refactor: import문 정리, type 폴더이동
