+++ src/app/admin/components/Button.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import { ButtonHTMLAttributes, ReactNode } from "react";
 
 interface FormButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
